LayoutTests/media/audio-play-with-video-element-expected.txt

Test that calling play() and pause() on a video element with audio source triggers play and pause events.


RUN(video.src = findMediaFile("audio", "content/silence"))
RUN(handlePromise(video.play()))
RUN(video.pause())
EVENT(play)
EVENT(waiting)
EVENT(pause)
EXPECTED (video.paused == 'true') OK
END OF TEST
